requirements.txt
aiohttp==3.9.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
scipy==1.11.4
soundfile==0.12.1
librosa==0.10.1
torch==2.1.2
torchaudio==2.1.2
deepgram-sdk==3.2.2
